pyproject.toml
[tool.poetry]
name = "ttools"
version = "0.1.0"
description = ""
authors = ["Askepios-ai <oyvind.a.nohr@uit.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
canvasapi = "^3.2.0"
pypdf = "^4.2.0"
textual = "^0.69.0"
fuse-python = "^1.0.8"
psutil = "^6.0.0"
pyte = "^0.8.2"
cachetools = "^5.3.3"
patool = "^2.3.0"


[tool.poetry.group.dev.dependencies]
textual-dev = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
